test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

hacking>=0.12.0,<0.13 # Apache-2.0

coverage>=4.0 # Apache-2.0
python-subunit>=0.0.18 # Apache-2.0/BSD
python-openstackclient>=3.3.0 # Apache-2.0
python-keystoneclient
testrepository>=0.0.18  # Apache-2.0/BSD
testscenarios>=0.4  # Apache-2.0/BSD
testtools>=1.4.0 # MIT
mock==2.0.0
